requirements.txt
Django>=5.1.11
djangorestframework>=3.14.0
Pillow>=10.0.0
